use extern const instead of #define to export symbols
Note this approach did not seem to work so well in that the
exported symbols were not the constant values, but indexes into
the HEAPU32 where the constant values could be found.

It appears that using EMSCRIPTEN_BINDINGS is a much cleaner
solution, though it is a little distasteful in that it
requires introducing a C++ source file into an otherwise
pure C codebase.